* casper/bootstrap: If an MBR was backed up before install started, restore it * ubiquity/bootstrap: If an MBR was backed up before install, don't restore DRMK   or syslinux MBR's.

+++ ubiquity/dell-bootstrap.py
@@ -1732,24 +1732,27 @@ class Install(InstallPlugin):
                     #we use the active partition bit in it
                     wfd.write('sfdisk -A%s %s\n' % (active, disk))
 
-                    #we don't necessarily know how we booted
-                    #test the md5 of the MBR to match DRMK or syslinux
-                    #if they don't match, rewrite MBR
-                    with misc.raised_privileges():
-                        with open(disk, 'rb') as rfd:
-                            disk_mbr = rfd.read(440)
-                    path = '/usr/share/dell/up/mbr.bin'
-                    if not os.path.exists(path):
-                        path = '/usr/lib/syslinux/mbr.bin'
-                    if not os.path.exists(path):
-                        raise RuntimeError, ("Missing DRMK and syslinux MBR")
-                    with open(path, 'rb') as rfd:
-                        file_mbr = rfd.read(440)
-                    if hashlib.md5(file_mbr).hexdigest() != hashlib.md5(disk_mbr).hexdigest():
-                        self.debug("%s: MBR of disk is invalid, rewriting" % NAME)
+                    #in factory process if we backed up an MBR, that would have already
+                    #been restored.
+                    if not os.path.exists(os.path.join(CDROM_MOUNT, 'factory', 'mbr.bin')):
+                        #we don't necessarily know how we booted
+                        #test the md5 of the MBR to match DRMK or syslinux
+                        #if they don't match, rewrite MBR
                         with misc.raised_privileges():
-                            with open(disk, 'wb') as wfd:
-                                wfd.write(file_mbr)
+                            with open(disk, 'rb') as rfd:
+                                disk_mbr = rfd.read(440)
+                        path = '/usr/share/dell/up/mbr.bin'
+                        if not os.path.exists(path):
+                            path = '/usr/lib/syslinux/mbr.bin'
+                        if not os.path.exists(path):
+                            raise RuntimeError, ("Missing DRMK and syslinux MBR")
+                        with open(path, 'rb') as rfd:
+                            file_mbr = rfd.read(440)
+                        if hashlib.md5(file_mbr).hexdigest() != hashlib.md5(disk_mbr).hexdigest():
+                            self.debug("%s: MBR of disk is invalid, rewriting" % NAME)
+                            with misc.raised_privileges():
+                                with open(disk, 'wb') as wfd:
+                                    wfd.write(file_mbr)
 
                 #If we have GPT, we need to go down other paths
                 elif layout == 'gpt':
